Refactor service namespaces to remove Impl subnamespace
